# public/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "POS System API",
    "description": "API for managing point of sale operations including products, customers, transactions, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@possystem.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/payments/subscription": {
      "post": {
        "summary": "Create subscription for membership products",
        "description": "Create Stripe subscriptions for membership products with automatic customer creation and billing setup",
        "tags": ["Payments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["cart"],
                "properties": {
                  "cart": {
                    "type": "object",
                    "description": "Shopping cart with membership products and connected customers",
                    "properties": {
                      "products": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string",
                              "description": "Product ID",
                              "example": "64f8a1b2c3d4e5f6a7b8c9d0"
                            },
                            "name": {
                              "type": "string",
                              "description": "Product name",
                              "example": "Premium Membership"
                            },
                            "type": {
                              "type": "string",
                              "enum": ["membership"],
                              "description": "Product type (must be membership)"
                            },
                            "variations": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string",
                                    "description": "Duration amount",
                                    "example": "1"
                                  },
                                  "unit": {
                                    "type": "string",
                                    "enum": ["month", "year", "week", "day", "fortnight"],
                                    "description": "Duration unit",
                                    "example": "month"
                                  },
                                  "prices": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "name": {
                                          "type": "string",
                                          "description": "Price tier name",
                                          "example": "Regular"
                                        },
                                        "value": {
                                          "type": "string",
                                          "description": "Price amount",
                                          "example": "29.99"
                                        },
                                        "customers": {
                                          "type": "array",
                                          "description": "Connected customers for this membership",
                                          "items": {
                                            "type": "object",
                                            "properties": {
                                              "customer": {
                                                "type": "object",
                                                "properties": {
                                                  "_id": {
                                                    "type": "string",
                                                    "description": "Customer ID",
                                                    "example": "64f8a1b2c3d4e5f6a7b8c9d1"
                                                  },
                                                  "name": {
                                                    "type": "string",
                                                    "description": "Customer name",
                                                    "example": "John Doe"
                                                  },
                                                  "email": {
                                                    "type": "string",
                                                    "description": "Customer email",
                                                    "example": "john@example.com"
                                                  },
                                                  "phone": {
                                                    "type": "string",
                                                    "description": "Customer phone",
                                                    "example": "+61400000000"
                                                  }
                                                }
                                              }
                                            }
                                          }
                                        }
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isSubscription": {
                      "type": "boolean",
                      "example": true
                    },
                    "transactionId": {
                      "type": "string",
                      "description": "Created transaction ID",
                      "example": "64f8a1b2c3d4e5f6a7b8c9d2"
                    },
                    "subscriptions": {
                      "type": "array",
                      "description": "Created Stripe subscriptions",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Stripe subscription ID",
                            "example": "sub_1234567890"
                          },
                          "status": {
                            "type": "string",
                            "description": "Subscription status",
                            "example": "active"
                          },
                          "productDetails": {
                            "type": "object",
                            "properties": {
                              "productId": {
                                "type": "string",
                                "description": "Product ID",
                                "example": "64f8a1b2c3d4e5f6a7b8c9d0"
                              },
                              "customerId": {
                                "type": "string",
                                "description": "Customer ID",
                                "example": "64f8a1b2c3d4e5f6a7b8c9d1"
                              },
                              "name": {
                                "type": "string",
                                "description": "Product name",
                                "example": "Premium Membership"
                              },
                              "variation": {
                                "type": "string",
                                "description": "Duration amount",
                                "example": "1"
                              },
                              "unit": {
                                "type": "string",
                                "description": "Duration unit",
                                "example": "month"
                              },
                              "price": {
                                "type": "string",
                                "description": "Price amount",
                                "example": "29.99"
                              }
                            }
                          }
                        }
                      }
                    },
                    "processedCustomers": {
                      "type": "array",
                      "description": "Customers processed with Stripe IDs",
                      "items": {
                        "type": "object",
                        "properties": {
                          "customerId": {
                            "type": "string",
                            "description": "Local customer ID",
                            "example": "64f8a1b2c3d4e5f6a7b8c9d1"
                          },
                          "stripeCustomerId": {
                            "type": "string",
                            "description": "Stripe customer ID",
                            "example": "cus_1234567890"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid cart data or no membership products"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error - Subscription creation failed"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Authentication token stored in HTTP-only cookie"
      }
    },
...
  },
...
}